Fixed batch updates observers bug
