+++ src/typings/index.ts
@@ -1,3 +1,4 @@
+export * from './BatchUpdate';
 export * from './Observer';
 export * from './MappingProxy';
 export * from './StockProxy';
