+++ test/hooks/useStock.test.ts
@@ -289,76 +289,3 @@ describe('Value setting and getting', () => {
         });
     });
 });
-
-describe('Batch observers tests', () => {
-    it('should call observer', () => {
-        const { result } = renderUseStockHook({
-            value: 'asdf',
-        });
-
-        const observer = jest.fn();
-
-        act(() => {
-            result.current.observeBatchUpdates(observer);
-
-            result.current.setValue('value', '2');
-        });
-
-        expect(observer).toBeCalled();
-    });
-
-    it('should call observer on any update with arguments', () => {
-        const { result } = renderUseStockHook({
-            value: 'asdf',
-            parent: {
-                child: {
-                    hello: 'asdf',
-                },
-            },
-        });
-
-        const observer = jest.fn();
-
-        act(() => {
-            result.current.observeBatchUpdates(observer);
-            result.current.observe('parent.child', jest.fn());
-            result.current.observe('parent', jest.fn());
-            result.current.observe('value', jest.fn());
-
-            result.current.setValue('parent.child.hello', '2');
-        });
-
-        expect(observer).toBeCalled();
-        expect(observer).toBeCalledWith({
-            paths: ['parent.child', 'parent'],
-            values: {
-                value: 'asdf',
-                parent: {
-                    child: {
-                        hello: '2',
-                    },
-                },
-            },
-        });
-    });
-
-    it('should remove observer', () => {
-        const { result } = renderUseStockHook({
-            value: 'asdf',
-        });
-
-        const observer = jest.fn();
-
-        act(() => {
-            const key = result.current.observeBatchUpdates(observer);
-
-            result.current.setValue('value', '2');
-
-            result.current.stopObservingBatchUpdates(key);
-
-            result.current.setValue('value', 'h');
-        });
-
-        expect(observer).toBeCalledTimes(1);
-    });
-});
